Update how unfiltered html is checked
