+++ code-block-pro.php
@@ -26,7 +26,6 @@ add_action('init', function () {
 add_action('admin_init', function () {
     wp_add_inline_script('kevinbatdorf-code-block-pro-editor-script', 'window.codeBlockPro = ' . wp_json_encode([
         'pluginUrl' => esc_url_raw(plugin_dir_url(__FILE__)),
-        'canSaveHtml' => current_user_can('unfiltered_html'),
     ]) . ';');
 });
