+++ cypress/e2e/permissions.cy.js
@@ -3,9 +3,6 @@ beforeEach(() => {
     cy.clearBrowserStorage();
     cy.loginUser();
     cy.visitNewPostEditor();
-    cy.window().then((win) => {
-        win.codeBlockPro.canSaveHtml = false;
-    });
 });
 afterEach(() => {
     cy.saveDraft(); // so we can leave without an alert
